changed from data-jdbc to jdbc

+++ src/test/java/ru/mtsbank/testproject/controllers/AccountsControllerTest.java
@@ -5,17 +5,14 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.HttpStatus;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
-import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import ru.mtsbank.testproject.services.AccountService;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @RunWith(SpringRunner.class)
 @WebMvcTest(AccountsController.class)
@@ -27,11 +24,11 @@ class AccountsControllerTest {
 
     @Test
     void createAccount() throws Exception {
-        Long id = 1L;
+        Integer id = 1;
         when(accountService.saveAccount(123, 1)).thenReturn(id);
         MvcResult result = mvc.perform(MockMvcRequestBuilders.post("/accounts").param("accountCurrency","123")
                 .param("clientId","1")).andReturn();
-        Long resultId = Long.valueOf(result.getResponse().getContentAsString());
+        Integer resultId = Integer.valueOf(result.getResponse().getContentAsString());
         assertEquals(id, resultId);
         System.out.println();
     }
